refactor: move /feed page to / index
